AC#663-3616 migrate private $options to protected $schedulerOptions
After the update to TYPO3 9 the private $options variable can't be
serialized and therefore not saved in the
database anymore as our parent implemented the __sleep() method to
return the class variables which should be
serialized/saved. So to keep the possibly saved $options we need to
move them to $schedulerOptions if present.
Otherwise the $options will be lost after the scheduler is executed/saved.

#11